requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
geopy>=2.3.0
python-dotenv>=1.0.0
shapely>=2.0.0
